Changed wind farm control calculation to be in-memory (instead of writing to file)

+++ Offwind.WebApp/Offwind.WebApp/Areas/EngineeringTools/Models/WindFarm/VWindFarm.cs
@@ -9,9 +9,10 @@ namespace Offwind.WebApp.Areas.EngineeringTools.Models.WindFarm
 {
     public sealed class VWindFarm : VWebPage
     {
-        public Double StartTime { set; get; }
-        public Double StopTime { set; get; }
-        public Double TimeStep { set; get; }
+        public decimal StartTime { set; get; }
+        public decimal StopTime { set; get; }
+        public decimal TimeStep { set; get; }
+
         [DisplayName("Wind farm scale")]
         public Double Scale { set; get; }
